feat: Domain層: GroupをUserから分離
