+++ api/user/internal/domain/repository/user.go
@@ -11,5 +11,4 @@ type UserRepository interface {
 	Authentication(ctx context.Context) (*domain.User, error)
 	Create(ctx context.Context, u *domain.User) error
 	GetUIDByEmail(ctx context.Context, email string) (string, error)
-	CreateGroup(ctx context.Context, u *domain.User) error
 }
