+++ api/user/internal/domain/service/user.go
@@ -15,7 +15,6 @@ import (
 type UserService interface {
 	Authentication(ctx context.Context) (*domain.User, error)
 	Create(ctx context.Context, u *domain.User) error
-	CreateGroup(ctx context.Context, u *domain.User, g *domain.Group) error
 }
 
 type userService struct {
@@ -48,22 +47,6 @@ func (us *userService) Create(ctx context.Context, u *domain.User) error {
 	}
 
 	if err := us.userRepository.Create(ctx, u); err != nil {
-		err = xerrors.Errorf("Failed to Domain/Repository: %w", err)
-		return domain.Unknown.New(err)
-	}
-
-	return nil
-}
-
-func (us *userService) CreateGroup(ctx context.Context, u *domain.User, g *domain.Group) error {
-	if err := us.userDomainValidation.Group(ctx, g); err != nil {
-		err = xerrors.Errorf("Failed to Domain/DomainValidation: %w", err)
-		return domain.InvalidDomainValidation.New(err)
-	}
-
-	u.Groups = append(u.Groups, *g)
-
-	if err := us.userRepository.CreateGroup(ctx, u); err != nil {
 		err = xerrors.Errorf("Failed to Domain/Repository: %w", err)
 		return domain.Unknown.New(err)
 	}
