+++ api/user/internal/domain/user.go
@@ -11,13 +11,5 @@ type User struct {
 	ThumbnailURL string    `firestore:"thumbnail_url"`
 	CreatedAt    time.Time `firestore:"created_at"`
 	UpdatedAt    time.Time `firestore:"updated_at"`
-	Groups       []Group   `firestore:"groups"`
-}
-
-// Group - Groupエンティティ
-type Group struct {
-	Name        string    `firestore:"name"`
-	Description string    `firestore:"description"`
-	CreatedAt   time.Time `firestore:"created_at"`
-	UpdatedAt   time.Time `firestore:"updated_at"`
+	Groups       []Group   `firestore:"-"`
 }
