+++ api/user/internal/domain/validation/group.go
@@ -1,0 +1,12 @@
+package validation
+
+import (
+	"context"
+
+	"github.com/16francs/gran/api/user/internal/domain"
+)
+
+// GroupDomainValidation - GroupDomainRepositoryインターフェース
+type GroupDomainValidation interface {
+	Group(ctx context.Context, g *domain.Group) error
+}
