+++ api/user/internal/domain/validation/user.go
@@ -9,5 +9,4 @@ import (
 // UserDomainValidation - UserDomainRepositoryインターフェース
 type UserDomainValidation interface {
 	User(ctx context.Context, u *domain.User) error
-	Group(ctx context.Context, g *domain.Group) error
 }
